requirements.txt
asgiref==3.5.2
Django==3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.8
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
django-environ
django-simple-history
django-admin-autocomplete-filter
Pillow
drf-yasg
django-import-export
jalali-django-admin-rangefilter
